pretrained/examples/ESPnet/inference/text_to_speech/requirements.txt
soundfile
espnet
espnet_model_zoo
